+++ src/test_util.rs
@@ -5,7 +5,7 @@ use std::sync::{Arc, OnceLock};
 use crate::ast::{Position, Span, Spanned};
 use crate::rust_span;
 
-pub fn test_span(start_line: usize, start_col: usize, end_line: usize, end_col: usize) -> Span {
+pub fn test_span(start_line: u32, start_col: u32, end_line: u32, end_col: u32) -> Span {
     Span::new(
         Position {
             offset: 0,
